GEODE-8624: Support Redis HINCRBYFLOAT command (#5986)
- Switch to using BigDecimal instead of double for increment value.
- Improve HINCRBYFLOAT output accuracy for very large values.
- Add concurrency tests.
- Mark HINCRBYFLOAT as supported.

+++ geode-redis/src/main/java/org/apache/geode/redis/internal/executor/CommandFunction.java
@@ -267,7 +267,7 @@ public class CommandFunction extends SingleResultRedisFunction {
       }
       case HINCRBYFLOAT: {
         ByteArrayWrapper field = (ByteArrayWrapper) args[1];
-        double increment = (double) args[2];
+        BigDecimal increment = (BigDecimal) args[2];
         return hashCommands.hincrbyfloat(key, field, increment);
       }
       default:
